--- lib/identity/hostdata/version.rb
@@ -1,5 +1,0 @@
-module Identity
-  module Hostdata
-    VERSION = "0.1.0"
-  end
-end
