+++ lib/login_gov/hostdata/version.rb
@@ -1,0 +1,5 @@
+module LoginGov
+  module Hostdata
+    VERSION = "0.2.0"
+  end
+end
